Working with the rofi install
This is tricky because on arch it's installed to the system, but on
Ubuntu I used the setup script that installs it to the user directory.

Similarly, on arch I can install rofi-git, where on Ubuntu I have to
clone it.

I'm attempting to set things up better here, but I still have to iron
out these conflicts.

That said, now screen shots would work. (also requires a tweak in the
rofi clone to specify the proper path for screen shots.  I'll have to
write my own launcher.)